requirements.txt
astroid==2.4.2
attrs==20.3.0
aws-cdk.assets==1.87.0
aws-cdk.aws-apigateway==1.87.0
aws-cdk.aws-apigatewayv2==1.87.0
aws-cdk.aws-applicationautoscaling==1.87.0
aws-cdk.aws-autoscaling==1.87.0
aws-cdk.aws-autoscaling-common==1.87.0
aws-cdk.aws-autoscaling-hooktargets==1.87.0
aws-cdk.aws-batch==1.87.0
aws-cdk.aws-certificatemanager==1.87.0
aws-cdk.aws-cloudformation==1.87.0
aws-cdk.aws-cloudfront==1.87.0
aws-cdk.aws-cloudwatch==1.87.0
aws-cdk.aws-codebuild==1.87.0
aws-cdk.aws-codecommit==1.87.0
aws-cdk.aws-codeguruprofiler==1.87.0
aws-cdk.aws-codepipeline==1.87.0
aws-cdk.aws-cognito==1.87.0
aws-cdk.aws-dynamodb==1.87.0
aws-cdk.aws-ec2==1.87.0
aws-cdk.aws-ecr==1.87.0
aws-cdk.aws-ecr-assets==1.87.0
aws-cdk.aws-ecs==1.87.0
aws-cdk.aws-efs==1.87.0
aws-cdk.aws-elasticloadbalancing==1.87.0
aws-cdk.aws-elasticloadbalancingv2==1.87.0
aws-cdk.aws-elasticsearch==1.87.0
aws-cdk.aws-events==1.87.0
aws-cdk.aws-events-targets==1.87.0
aws-cdk.aws-iam==1.87.0
aws-cdk.aws-kinesis==1.87.0
aws-cdk.aws-kinesisfirehose==1.87.0
aws-cdk.aws-kms==1.87.0
aws-cdk.aws-lambda==1.87.0
aws-cdk.aws-logs==1.87.0
aws-cdk.aws-route53==1.87.0
aws-cdk.aws-route53-targets==1.87.0
aws-cdk.aws-s3==1.87.0
aws-cdk.aws-s3-assets==1.87.0
aws-cdk.aws-sam==1.87.0
aws-cdk.aws-secretsmanager==1.87.0
aws-cdk.aws-servicediscovery==1.87.0
aws-cdk.aws-sns==1.87.0
aws-cdk.aws-sns-subscriptions==1.87.0
aws-cdk.aws-sqs==1.87.0
aws-cdk.aws-ssm==1.87.0
aws-cdk.aws-stepfunctions==1.87.0
aws-cdk.cloud-assembly-schema==1.87.0
aws-cdk.core==1.87.0
aws-cdk.custom-resources==1.87.0
aws-cdk.cx-api==1.87.0
aws-cdk.region-info==1.87.0
aws-lambda==2.1.3
boto3==1.16.25
botocore==1.19.25
cattrs==1.1.1
cfn-flip==1.2.3
cfnresponse==1.1.1
click==7.1.2
constructs==3.2.47
isort==5.6.4
jmespath==0.10.0
jsii==1.15.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
publication==0.0.3
pylint==2.6.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
toml==0.10.2
troposphere==2.6.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
